requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
mysqlclient>=2.2
whitenoise>=6.6
gunicorn>=22.0
python-dotenv>=1.0
Pillow>=10.0
qrcode>=7.4,<8.0
django-cors-headers==4.4.0
PyMySQL==1.1.0